requirements.txt
asgiref==3.6.0
Django==4.1.8
djangorestframework==3.14.0
mssql-django==1.2
pyodbc==4.0.38
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
